backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
lxml
cohere
groq
python-dotenv
